Introduction
I want to work in big companies and I have a passion for the work. Currently, I'm learning cloud computing. I'm a bookaholic too... I read a lot of books and I write books. Watching movies. But most of the time I tried to learn new things and solve problems and research new things and technologies.